# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestion Social ONG
  description: API generica para consultar las tablas del sistema de beneficiarios, donaciones, inventario, misiones y entregas.
  version: 3.0.0

servers:
  - url: http://localhost:3000
    description: Desarrollo local

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [usuario, contraseña]
      properties:
        usuario:
          type: string
          example: admin
        contraseña:
          type: string
          example: admin123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          type: object
          properties:
            id:
              type: integer
            nombre:
              type: string
            correo:
              type: string
            rol:
              type: string
    SearchRequest:
      type: object
      required: [primaryTable]
      properties:
        primaryTable:
          type: string
          example: beneficiario
        filters:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: numero_documento
              operator:
                type: string
                example: ILIKE
              value:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: array
        orderBy:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: fecha_registro
              direction:
                type: string
                example: DESC

security:
  - BearerAuth: []

paths:
  /api/health:
    get:
      security: []
      summary: Verificar estado del backend y conexion a base de datos
      responses:
        '200':
          description: Servicio activo
        '500':
          description: Error de conexion o servidor

  /api/auth/login:
    post:
      security: []
      summary: Iniciar sesion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Acceso concedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales invalidas

  /api/auth/profile:
    get:
      summary: Obtener perfil del usuario autenticado
      responses:
        '200':
          description: Perfil del usuario autenticado
        '401':
          description: Token no proporcionado
        '403':
          description: Token invalido o expirado

  /api/meta/app:
    get:
      summary: Obtener metadatos del aplicativo
      description: Retorna entidades permitidas, columnas, llaves primarias y totales por tabla.
      responses:
        '200':
          description: Metadatos cargados

  /api/records/{table}:
    get:
      summary: Listar registros paginados de una tabla permitida
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
          example: beneficiario
        - in: query
          name: page
          schema:
            type: integer
          example: 1
        - in: query
          name: pageSize
          schema:
            type: integer
          example: 15
        - in: query
          name: q
          schema:
            type: string
          example: Perez
        - in: query
          name: sortField
          schema:
            type: string
          example: id
        - in: query
          name: sortDirection
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: Registros paginados
        '404':
          description: Tabla no permitida
    post:
      summary: Crear registro
      description: Actualmente deshabilitado porque la aplicacion esta en modo solo consulta.
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      responses:
        '403':
          description: Modo solo consulta

  /api/records/{table}/{id}:
    get:
      summary: Obtener registro por ID en tablas con llave primaria simple
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
          example: beneficiario
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: '1'
      responses:
        '200':
          description: Registro encontrado
        '404':
          description: Registro no encontrado
    put:
      summary: Actualizar registro
      description: Actualmente deshabilitado porque la aplicacion esta en modo solo consulta.
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '403':
          description: Modo solo consulta
    delete:
      summary: Eliminar registro
      description: Actualmente deshabilitado para llaves simples porque la aplicacion esta en modo solo consulta.
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '403':
          description: Modo solo consulta

  /api/search/execute:
    post:
      summary: Ejecutar busqueda avanzada parametrizada
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          example: 1
        - in: query
          name: pageSize
          schema:
            type: integer
          example: 20
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Resultados de busqueda
        '400':
          description: Parametros invalidos

  /api/search/schema:
    get:
      summary: Obtener esquema filtrado de tablas permitidas
      responses:
        '200':
          description: Esquema disponible

  /api/search/tables:
    get:
      summary: Obtener lista de tablas permitidas
      responses:
        '200':
          description: Lista de tablas